+++ packages/isaacscript-common/src/cachedClasses.ts
@@ -3,7 +3,7 @@
  *
  * Use this instead of invoking the constructor again for a miniscule performance increase.
  *
- * Caching the results of a this constructor is safe, but caching other classes (like `Level` or
+ * Caching the results of this constructor is safe, but caching other classes (like `Level` or
  * `Room`) is not safe and can lead to the game crashing in certain situations.
  */
 export const game = Game();
@@ -13,7 +13,7 @@ export const game = Game();
  *
  * Use this instead of invoking the constructor again for a miniscule performance increase.
  *
- * Caching the results of a this constructor is safe, but caching other classes (like `Level` or
+ * Caching the results of this constructor is safe, but caching other classes (like `Level` or
  * `Room`) is not safe and can lead to the game crashing in certain situations.
  */
 export const itemConfig = Isaac.GetItemConfig();
@@ -23,7 +23,7 @@ export const itemConfig = Isaac.GetItemConfig();
  *
  * Use this instead of invoking the constructor again for a miniscule performance increase.
  *
- * Caching the results of a this constructor is safe, but caching other classes (like `Level` or
+ * Caching the results of this constructor is safe, but caching other classes (like `Level` or
  * `Room`) is not safe and can lead to the game crashing in certain situations.
  */
 export const musicManager = MusicManager();
@@ -33,7 +33,7 @@ export const musicManager = MusicManager();
  *
  * Use this instead of invoking the constructor again for a miniscule performance increase.
  *
- * Caching the results of a this constructor is safe, but caching other classes (like `Level` or
+ * Caching the results of this constructor is safe, but caching other classes (like `Level` or
  * `Room`) is not safe and can lead to the game crashing in certain situations.
  */
 export const sfxManager = SFXManager();