+++ packages/isaacscript-common/src/features/saveDataManager/load.ts
@@ -1,5 +1,6 @@
 import { jsonDecode } from "../../functions/jsonHelpers";
 import { log, logError } from "../../functions/log";
+import { iterateTableDeterministically } from "../../functions/table";
 import { SaveData } from "../../interfaces/SaveData";
 import {
   SAVE_DATA_MANAGER_DEBUG,
@@ -27,35 +28,39 @@ export function loadFromDisk(
   }
 
   // Second, iterate over all the fields of the new table.
-  for (const [key, value] of pairs(newSaveData)) {
-    // All elements of loaded save data should have keys that are strings equal to the name of the
-    // subscriber/feature. Ignore elements with other types of keys.
-    if (typeof key !== "string") {
-      continue;
-    }
+  iterateTableDeterministically(
+    newSaveData,
+    (key, value) => {
+      // All elements of loaded save data should have keys that are strings equal to the name of the
+      // subscriber/feature. Ignore elements with other types of keys.
+      if (typeof key !== "string") {
+        return;
+      }
 
-    // All elements of loaded save data should be tables that contain fields corresponding to the
-    // SaveData interface. Ignore elements that are not tables.
-    const valueType = type(value);
-    if (valueType !== "table") {
-      continue;
-    }
+      // All elements of loaded save data should be tables that contain fields corresponding to the
+      // SaveData interface. Ignore elements that are not tables.
+      const valueType = type(value);
+      if (valueType !== "table") {
+        return;
+      }
 
-    // Ignore elements that represent subscriptions that no longer exist in the current save data.
-    const oldSaveDataForSubscriber = oldSaveData.get(key);
-    if (oldSaveDataForSubscriber === undefined) {
-      continue;
-    }
+      // Ignore elements that represent subscriptions that no longer exist in the current save data.
+      const oldSaveDataForSubscriber = oldSaveData.get(key);
+      if (oldSaveDataForSubscriber === undefined) {
+        return;
+      }
 
-    if (SAVE_DATA_MANAGER_DEBUG) {
-      log(`Merging in stored data for feature: ${key}`);
-    }
+      if (SAVE_DATA_MANAGER_DEBUG) {
+        log(`Merging in stored data for feature: ${key}`);
+      }
 
-    // We do not want to blow away the child tables of the existing map, because save data could
-    // contain out-of-date fields. Instead, merge it one field at a time in a recursive way (and
-    // convert Lua tables back to TypeScriptToLua Maps, if necessary).
-    merge(oldSaveDataForSubscriber as LuaTable, value as LuaTable, key);
-  }
+      // We do not want to blow away the child tables of the existing map, because save data could
+      // contain out-of-date fields. Instead, merge it one field at a time in a recursive way (and
+      // convert Lua tables back to TypeScriptToLua Maps, if necessary).
+      merge(oldSaveDataForSubscriber as LuaTable, value as LuaTable, key);
+    },
+    SAVE_DATA_MANAGER_DEBUG,
+  );
 
   log(
     `The ${SAVE_DATA_MANAGER_FEATURE_NAME} loaded data from the "save#.dat" file.`,